+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.2.0] - 2026-09-06
+
 - Only show minor aspects if the regular aspect has been unlocked.
 - Update random fusion to account for unlocked aspects.
 - Ensure that `Aspect of Melinoe` for every weapon is always considered unlocked.
@@ -198,7 +200,8 @@
 
 - First version of the mod!
 
-[unreleased]: https://github.com/adi1998/WeaponFusion/compare/1.1.3...HEAD
+[unreleased]: https://github.com/adi1998/WeaponFusion/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/adi1998/WeaponFusion/compare/1.1.3...1.2.0
 [1.1.3]: https://github.com/adi1998/WeaponFusion/compare/1.1.2...1.1.3
 [1.1.2]: https://github.com/adi1998/WeaponFusion/compare/1.1.1...1.1.2
 [1.1.1]: https://github.com/adi1998/WeaponFusion/compare/1.1.0...1.1.1